requirements.txt
pyspark
findspark
kafka-python
faker
python-dotenv